bot/requirements.txt
discord.py==1.3.4
requests==2.21.0
numpy==1.16.2
aiohttp==3.6.2
gspread==3.6.0
pymongo[tls,srv]==3.10.1
pytz==2018.9
discord==1.0.1
python_dateutil==2.8.1
Pillow==7.2.0